Remove clip hull generation system
Deletes bsp_hull.h/cpp, test_hull_csg.cpp, and docs/clip-hull-collision.md.
Removes build_debug_hull_meshes(), classify_point_hull(), and all related
clip hull pipeline code from goldsrc_bsp.cpp. The minimal walk_bsp_tree and
compute_cell_vertices helpers needed by build_brush_convex (brush entity and
water collision) are inlined directly into goldsrc_bsp.cpp.
